GITOPSRVCE-792: Light refactoring of DT/DTC/DTClaim controllrs
